fix: Move SchemaUtils and SchemaResolver to connector from optimizer
