+++ axiom/connectors/tests/CMakeLists.txt
@@ -16,7 +16,7 @@ add_library(axiom_test_connector TestConnector.cpp)
 
 target_link_libraries(
   axiom_test_connector
-  axiom_connector_metadata
+  axiom_connectors
   velox_common_base
   velox_memory
   velox_connector
@@ -37,3 +37,9 @@ target_link_libraries(
   gtest
   gtest_main
 )
+
+add_executable(axiom_connectors_tests SchemaUtilsTest.cpp SchemaResolverTest.cpp)
+
+add_test(axiom_connectors_tests axiom_connectors_tests)
+
+target_link_libraries(axiom_connectors_tests axiom_connectors axiom_test_connector gtest gtest_main)
