+++ axiom/connectors/tpch/CMakeLists.txt
@@ -16,7 +16,7 @@ add_library(axiom_tpch_connector_metadata OBJECT TpchConnectorMetadata.cpp)
 
 target_link_libraries(
   axiom_tpch_connector_metadata
-  axiom_connector_metadata
+  axiom_connectors
   velox_tpch_connector
   velox_tpch_gen
 )
