+++ axiom/logical_plan/CMakeLists.txt
@@ -28,7 +28,7 @@ add_library(axiom_logical_plan_builder PlanBuilder.cpp NameAllocator.cpp NameMap
 target_link_libraries(
   axiom_logical_plan_builder
   axiom_logical_plan
-  axiom_connector_metadata
+  axiom_connectors
   velox_exec
   velox_function_registry
   velox_functions_prestosql
