+++ axiom/runner/CMakeLists.txt
@@ -25,7 +25,7 @@ add_library(axiom_runner_local_runner LocalRunner.cpp Runner.cpp)
 target_link_libraries(
   axiom_runner_local_runner
   axiom_runner_multifragment_plan
-  axiom_connector_metadata
+  axiom_connectors
   velox_common_base
   velox_memory
   velox_hive_connector
